// src/types/slices/productsOrderingPageSliceType.d.ts
export interface productsOrderingPageSliceType {
    clientName: string,
    phoneNumber: string,
    addressEmail: string,
    address: string,
    zipCode: string,
    city: string,
    additionalInformation: string,
    typeOfDelivery: string,
    dateOfScheduledDeliveryHome: string,
    timeOfScheduledDeliveryHome: string,
    dateOfScheduledDeliveryPickup: string,
    timeOfScheduledDeliveryPickup: string,
    typeOfPayment: string,
    agreement: boolean
}